requirements.txt
convertdate==2.1.3
cycler==0.10.0
Cython==0.29.2
ephem==3.7.6.0
fbprophet==0.4
holidays==0.9.8
kiwisolver==1.0.1
lunardate==0.2.0
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pyparsing==2.3.0
pystan==2.18.1.0
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
stopwatch.py==1.0.1
